Remove `Scheduler` and move items to `Executor`
